fix: fund arbitrum
